fix messages
