+++ app/src/components/ChatDisplay.svelte
@@ -7,7 +7,7 @@
     import { setEmoteSize, settings } from "$stores/settings";
     import { badges, globals } from "$stores/global";
 
-    let chatMessages: Record<string, any>[] = [];
+    let chatMessages: Record<string, any>[] = $state([]);
     let chat: HTMLElement;
 
     let styles: Record<string, string | number> = {};
@@ -232,6 +232,56 @@
                 };
             }) as any,
     );
+
+    function randomString(len: number) {
+        const chars = "abcdefghijklmnopqrstuvwxyz0123456789";
+        let out = "";
+        for (let i = 0; i < len; i++)
+            out += chars[Math.floor(Math.random() * chars.length)];
+        return out;
+    }
+
+    function pickRandomBadges(): Record<string, string>[] {
+        const count = Math.floor(Math.random() * 2) + 1;
+        const shuffled = $badges["TTV"]["global"].sort(
+            () => 0.5 - Math.random(),
+        );
+        return shuffled.slice(0, count);
+    }
+
+    function sendFakeMessage(message: string) {
+        const username = randomString(5);
+        const displayName = username;
+        const userId = Math.floor(Math.random() * 1_000_000_000).toString();
+
+        const badgesPicked = pickRandomBadges() as Record<string, string>[];
+
+        const badgesRaw = badgesPicked
+            .map((b: Record<string, string>) => {
+                const badge_split = b.id.split("_");
+
+                return `${badge_split[0]}/${badge_split[1]}`;
+            })
+            .join(",");
+        const badges_parsed: Record<string, string> = {};
+        badgesPicked.forEach((b: Record<string, string>) => {
+            const badge_split = b.id.split("_");
+
+            badges_parsed[badge_split[0]] = badge_split[1];
+        });
+
+        const tags = {
+            username,
+            "display-name": displayName,
+            "user-id": userId,
+            "badges-raw": badgesRaw,
+            badges,
+            color: null,
+            "room-id": "0",
+        };
+
+        messages.update((msgs) => [...msgs, { tags, message }]);
+    }
 
     onMount(() => {
         let lastScrollTime = 0;
